Move utils to root
