+++ commands/trades.py
@@ -5,7 +5,7 @@ from gql import gql
 
 from constants import SEPARATOR, COLOR_LABEL, COLOR_LABEL_DELETED, COLOR_SEPARATOR
 from commands.tokens import TOKEN_FIELDS_BASIC, to_token
-from helpers.utils import get_graphql_client, to_etherscan_link, format_integer, format_amount_in_weis, format_token_long, format_token_short, debug_query, format_date_time
+from utils import get_graphql_client, to_etherscan_link, format_integer, format_amount_in_weis, format_token_long, format_token_short, debug_query, format_date_time
 
 # Trade entity fields
 #   See https://thegraph.com/explorer/subgraph/gnosis/dfusion
